# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blood Donor Application API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
  database?: string;
  error?: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Donor {
  id: number;
  name: string;
  bloodType: string;
  phone: string;
  email: string;
  city: string;
  address?: string;
  isAvailable: boolean;
  lastDonationDate?: string | null;
  createdAt: string;
}

export interface CreateDonorRequest {
  name: string;
  bloodType: string;
  phone: string;
  email: string;
  city: string;
  address?: string;
  lastDonationDate?: string | null;
}

export interface UpdateDonorRequest {
  isAvailable?: boolean;
  lastDonationDate?: string | null;
  phone?: string;
  city?: string;
  address?: string;
}

export interface BloodRequest {
  id: number;
  patientName: string;
  bloodType: string;
  unitsNeeded: number;
  hospital: string;
  city: string;
  urgency: string;
  status: string;
  contactPhone: string;
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateBloodRequestBodyUrgency =
  (typeof CreateBloodRequestBodyUrgency)[keyof typeof CreateBloodRequestBodyUrgency];

export const CreateBloodRequestBodyUrgency = {
  critical: "critical",
  urgent: "urgent",
  normal: "normal",
} as const;

export interface CreateBloodRequestBody {
  patientName: string;
  bloodType: string;
  unitsNeeded: number;
  hospital: string;
  city: string;
  urgency: CreateBloodRequestBodyUrgency;
  contactPhone: string;
  notes?: string | null;
}

export type UpdateBloodRequestBodyStatus =
  (typeof UpdateBloodRequestBodyStatus)[keyof typeof UpdateBloodRequestBodyStatus];

export const UpdateBloodRequestBodyStatus = {
  open: "open",
  fulfilled: "fulfilled",
  cancelled: "cancelled",
} as const;

export interface UpdateBloodRequestBody {
  status?: UpdateBloodRequestBodyStatus;
  unitsNeeded?: number;
}

export interface Notification {
  id: number;
  donorId: number;
  bloodRequestId: number;
  message: string;
  isRead: boolean;
  createdAt: string;
}

export interface BloodTypeStat {
  /** The blood type (A+, A-, B+, B-, AB+, AB-, O+, O-) */
  bloodType: string;
  donorCount: number;
  requestCount: number;
}

export interface DashboardStats {
  totalDonors: number;
  availableDonors: number;
  totalRequests: number;
  openRequests: number;
  criticalRequests: number;
  bloodTypeStats: BloodTypeStat[];
}

export type ListDonorsParams = {
  bloodType?: string;
  city?: string;
  available?: boolean;
};

export type ListBloodRequestsParams = {
  bloodType?: string;
  urgency?: string;
  status?: string;
};

export type ListNotificationsParams = {
  donorId: number;
  unread?: boolean;
};