# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blood Donor Application API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
  database: zod.string().optional(),
  error: zod.string().optional(),
});

/**
 * @summary List all donors
 */
export const ListDonorsQueryParams = zod.object({
  bloodType: zod.coerce.string().optional(),
  city: zod.coerce.string().optional(),
  available: zod.coerce.boolean().optional(),
});

export const ListDonorsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  bloodType: zod.string(),
  phone: zod.string(),
  email: zod.string(),
  city: zod.string(),
  address: zod.string().optional(),
  isAvailable: zod.boolean(),
  lastDonationDate: zod.string().nullish(),
  createdAt: zod.string(),
});
export const ListDonorsResponse = zod.array(ListDonorsResponseItem);

/**
 * @summary Register a new donor
 */
export const CreateDonorBody = zod.object({
  name: zod.string(),
  bloodType: zod.string(),
  phone: zod.string(),
  email: zod.string(),
  city: zod.string(),
  address: zod.string().optional(),
  lastDonationDate: zod.string().nullish(),
});

/**
 * @summary Get donor by ID
 */
export const GetDonorParams = zod.object({
  id: zod.coerce.number(),
});

export const GetDonorResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  bloodType: zod.string(),
  phone: zod.string(),
  email: zod.string(),
  city: zod.string(),
  address: zod.string().optional(),
  isAvailable: zod.boolean(),
  lastDonationDate: zod.string().nullish(),
  createdAt: zod.string(),
});

/**
 * @summary Update donor availability
 */
export const UpdateDonorParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateDonorBody = zod.object({
  isAvailable: zod.boolean().optional(),
  lastDonationDate: zod.string().nullish(),
  phone: zod.string().optional(),
  city: zod.string().optional(),
  address: zod.string().optional(),
});

export const UpdateDonorResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  bloodType: zod.string(),
  phone: zod.string(),
  email: zod.string(),
  city: zod.string(),
  address: zod.string().optional(),
  isAvailable: zod.boolean(),
  lastDonationDate: zod.string().nullish(),
  createdAt: zod.string(),
});

/**
 * @summary List all blood requests
 */
export const ListBloodRequestsQueryParams = zod.object({
  bloodType: zod.coerce.string().optional(),
  urgency: zod.coerce.string().optional(),
  status: zod.coerce.string().optional(),
});

export const ListBloodRequestsResponseItem = zod.object({
  id: zod.number(),
  patientName: zod.string(),
  bloodType: zod.string(),
  unitsNeeded: zod.number(),
  hospital: zod.string(),
  city: zod.string(),
  urgency: zod.string(),
  status: zod.string(),
  contactPhone: zod.string(),
  notes: zod.string().nullish(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const ListBloodRequestsResponse = zod.array(
  ListBloodRequestsResponseItem,
);

/**
 * @summary Create a blood request
 */
export const CreateBloodRequestBody = zod.object({
  patientName: zod.string(),
  bloodType: zod.string(),
  unitsNeeded: zod.number(),
  hospital: zod.string(),
  city: zod.string(),
  urgency: zod.enum(["critical", "urgent", "normal"]),
  contactPhone: zod.string(),
  notes: zod.string().nullish(),
});

/**
 * @summary Get blood request by ID
 */
export const GetBloodRequestParams = zod.object({
  id: zod.coerce.number(),
});

export const GetBloodRequestResponse = zod.object({
  id: zod.number(),
  patientName: zod.string(),
  bloodType: zod.string(),
  unitsNeeded: zod.number(),
  hospital: zod.string(),
  city: zod.string(),
  urgency: zod.string(),
  status: zod.string(),
  contactPhone: zod.string(),
  notes: zod.string().nullish(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Update blood request status
 */
export const UpdateBloodRequestParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateBloodRequestBody = zod.object({
  status: zod.enum(["open", "fulfilled", "cancelled"]).optional(),
  unitsNeeded: zod.number().optional(),
});

export const UpdateBloodRequestResponse = zod.object({
  id: zod.number(),
  patientName: zod.string(),
  bloodType: zod.string(),
  unitsNeeded: zod.number(),
  hospital: zod.string(),
  city: zod.string(),
  urgency: zod.string(),
  status: zod.string(),
  contactPhone: zod.string(),
  notes: zod.string().nullish(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary List notifications for a donor
 */
export const ListNotificationsQueryParams = zod.object({
  donorId: zod.coerce.number(),
  unread: zod.coerce.boolean().optional(),
});

export const ListNotificationsResponseItem = zod.object({
  id: zod.number(),
  donorId: zod.number(),
  bloodRequestId: zod.number(),
  message: zod.string(),
  isRead: zod.boolean(),
  createdAt: zod.string(),
});
export const ListNotificationsResponse = zod.array(
  ListNotificationsResponseItem,
);

/**
 * @summary Mark notification as read
 */
export const MarkNotificationReadParams = zod.object({
  id: zod.coerce.number(),
});

export const MarkNotificationReadResponse = zod.object({
  id: zod.number(),
  donorId: zod.number(),
  bloodRequestId: zod.number(),
  message: zod.string(),
  isRead: zod.boolean(),
  createdAt: zod.string(),
});

/**
 * @summary Get dashboard statistics
 */
export const GetStatsResponse = zod.object({
  totalDonors: zod.number(),
  availableDonors: zod.number(),
  totalRequests: zod.number(),
  openRequests: zod.number(),
  criticalRequests: zod.number(),
  bloodTypeStats: zod.array(
    zod.object({
      bloodType: zod
        .string()
        .describe("The blood type (A+, A-, B+, B-, AB+, AB-, O+, O-)"),
      donorCount: zod.number(),
      requestCount: zod.number(),
    }),
  ),
});