# lib/api-zod/src/generated/types/bloodTypeStat.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blood Donor Application API
 * OpenAPI spec version: 0.1.0
 */

export interface BloodTypeStat {
  /** The blood type (A+, A-, B+, B-, AB+, AB-, O+, O-) */
  bloodType: string;
  donorCount: number;
  requestCount: number;
}